requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
cffi==1.15.0
cryptography==36.0.1
Django==4.0
djangorestframework==3.13.1
mysql-client==0.0.1
mysqlclient==2.1.0
numpy==1.21.4
pandas==1.3.5
pycparser==2.21
PyMySQL==1.0.2
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
